requirements.txt
selenium==3.141.0
webdriver-manager==3.2.2
pyyaml==5.3.1
pytest==6.1.2
pytest-html==2.1.1